Remove memberof attributes from group's UID listing
When retrieving users_for_gid for an AD group that contains another,
empty group as a member, this prevents an infinite recursion.

Previously users_for_gid would be called against the empty group.  Since
it had no members, `search.respond_to?(:member)` was false and the
method returned the result of `search.memberof`.  This caused the member
lookup to check the outer group again, resulting in an infinite loop.